rename: 폴더구조 컨벤션에 맞추어서 변경 (#111)
