+++ src/pages/search/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import SearchBar from '@/pages/search/SearchBar';
-import TabContainer from '@/pages/search/TabContainer';
+import SearchBar from '@/pages/search/components/SearchBar';
+import TabContainer from '@/pages/search/components/TabContainer';
 import Flex from '@/components/Flex';
 import Header from '@/components/Header';
-import { DEFAULT_SORT_TAGS } from '@/constants/defaultSortTags';
+import { DEFAULT_SORT_TAGS } from '@/pages/search/constants/DefaultSortTags';
 import { headerRootCutomStyle } from './index.css';
 
 const Search = () => {
